Cleanup warnings with MSVC 2003
git-svn-id: https://svn.pjsip.org/repos/pjproject/trunk@472 74dad513-b988-da41-8d7b-12977e46ad98

+++ pjlib-util/include/pjlib-util/scanner.h
@@ -513,7 +513,7 @@ PJ_DECL(void) pj_scan_restore_state( pj_scanner *scanner,
  */
 PJ_INLINE(int) pj_scan_get_col( const pj_scanner *scanner )
 {
-    return scanner->curptr - scanner->start_line;
+    return (int)(scanner->curptr - scanner->start_line);
 }
 
 /**
